issue-998: extract tool policy and tools crate
